Assembly Analysis Flows/Taks/Models - orchestration
This is a bit of departure from the current mode of working for
raw-reads and amplicon. I've created models and elements in the code
to keep track of the Assembly Analysis Pipeline (ASA), VIRIfy and
Mobilome Annotation Pipeline (MAP) with bespoke django models and code.

I'm not saying this is the best way, but I think our team will be able to
more easly work with this model... let's see.

I've addeda AssemblyAnalysisBatch set of models, each time the main
analysis_assembly_study is trigger it will chunk the assemblies to
analize in batches. These batches are stored in the database, and it's
possible to disable individual batch-analyses (we will add more features
to manage them if this works).

The code still needs some cleaning as I was trying to get a working version
before the end of the sprint. I've added unit tests and got a decent coverage
of the flows and tasks. But until we run this with real data we won't know
how it works.

Again, this is massive PR apologies again.