Another big refactor for ASA-MAP-Virify
These changes are an attempt to make is easier to restart batches from
partial fails (subsets of pipeline executions failling at different points).

I still think some more thought needs to be put into this, this implementation is
very "imperative" but I think a more hook based one should be much easier to maintain.

I've removed the AssemblyAnalysisBatch status per pipeline with a count based view.
If a batch fails or not is not binary, the user will have to kind-of-decide based
on the number of batch jobs that fail at different pipelines.

I've also added a batch job log collection thingy, hopefully this will make it easier to debug. This
thing is build with JsonField (_a la Sandy_).

Claude massively helped me write this, I read all the code and extensively tweaked it
but I was in a bit of a rush so I may be forgotten to remove some robot-non-sense (apologies for that).