Close old connections for the analysis flows
I recently bumped into a flow that took a couple of days to
complete (some soil assemblies analysis). When ASA was completed
and it tried to update the DB in the "main" flow the DB connection
was closed. This is to prevent that, as django doesn't automatically
handles that outside the request cycle.

We need a more elegant way of doing this, probably with out own signal of
sorts. I'm defering that for now